Switched the backend deatabase from SQLite to PostgreSQL
Also merged in some error handling I created on another computer to handle issues related to not choosing any indicators.

This change required reworking how database queries were created. In doing so I streamlined some of the database code making it easier. Also improved the CSV code to make it more robust to changes and less dependent on the to string method for database entries.